add integration test coverage for core
Core integration tests send string commands and perform rudimentary
asserts on the output, insofar as is possible with the random generator.
More fine-grained testing will be left for unit tests. Eventually, once
it becomes possible to seed data, we'll be able to be a bit more clever
with our asserts.