switch tests from StepRng to SmallRng
StepRng was a pain to get random-looking data from, and required the use
of traits at levels where traits really weren't necessary. Switched all
tests to use SmallRng with a hard-coded seed instead, matching the
application logic that uses SmallRng with a random seed.

(This may break in the future if SmallRng changes algorithms, but we can
cross that bridge when we get to it.)