impl saving to database
This is it, folks. We can now dump data into the black hole that is the
data store. Currently no way to read it back, however.

Removing the WASM tests for now. They're not providing a whole lot of
value, and would probably be better handled with something
Selenium-based that can cover the JS side of things as well.