fix things being silently dropped on failed save
The save operation would first remove a Thing from the recent list,
then attempt to save it. Fixed so that a failed operation causes the
Thing to be re-added to the end of the recent list. A bit clunky since
the order of eviction is lost, but this is an edge case.