add tests for CommandAlias
I got lazy and never wrote tests for the CommandAlias struct on the
excuse that it was a skinny wrapper for other classes and a bit hard to
unit test. However, with increased complexity comes increased need to
test, so time to fill that hole before this gets out of hand.