support Result<T, String> when deriving newtypes
In addition to deriving Display and ContextAwareParse on
`NewType(Vec<T>)`, it can now be derived on
`NewType(Vec<Result<T, String>>)`. This allows unsupported terms to be
bubbled up to the calling code. Unfortunately, the positional
information is lost in the intermediate layers, but hopefully that's not
too much of a problem at this point.