Display autocomplete suggestion list
The entire list of autocomplete suggestions is now displayed in a menu
similar to the one used by the web front end.

Needed to restructure the drawing some to avoid overlap and get things
updating correctly as the user types.

Still need to implement navigating up and down the list of selections.