introduce (Thing|Npc|Place)Data
We still need a way to represent a blob of data that is not a complete
Thing, for the purposes of undo/redo, (re)generating information, etc.
This is now done by following the pattern of Thing { uuid, data }, where
all meta fields are contained within data. (Parallels of this pattern
have also been created specifically for Npc and Place.)

To keep this change "small", the `uuid` property of `Thing` remains
`Option`al; I'll tighten up the distinction between `Thing` and
`ThingData` in a future commit.