refactor out is_saved flag for Things
Boolean flag was used to indicate if a Thing was currently located in
the journal or temporarily stored in recent entries. This wasn't ideal
because the data saved to journal/recent included the flag, so there was
a lot of bespoke logic to ensure (maybe unsuccessfully) that the flag
was never accidentally saved to the wrong place.

Instead, we wrap the Thing in an ephemeral wrapper called Record,
making the saved-ness of the Thing accessible through the return value
if needed but without bothering to include it in either the stored data
or code that interacts with the data directly without regard for where
it came from.